Move source to src folder in structure, sh files to scripts, enhance to use new path
