+++ .vscode/launch.json
@@ -5,7 +5,7 @@
             "type": "node",
             "request": "launch",
             "name": "Server Debug",
-            "program": "${workspaceFolder}/server/server.js",
+            "program": "${workspaceFolder}/server/src/server.js",
             "preLaunchTask": "decryptCredsForLocalRun",
             "envFile": "${workspaceFolder}/.env"
         }
