+++ .vscode/tasks.json
@@ -2,7 +2,7 @@
     "version": "2.0.0",
     "tasks": [{
         "label": "decryptCredsForLocalRun",
-        "command": "AWS_PROFILE=peblio ./server/export_creds.sh",
+        "command": "AWS_PROFILE=peblio ./server/scripts/export_creds.sh",
         "type": "shell"
     }]
 }
