+++ server/scripts/add_credstash_secret.sh
