+++ server/scripts/delete_credstash_secret.sh
