+++ server/scripts/export_creds.sh
