+++ server/scripts/get_credstash_secret.sh
