+++ server/scripts/list_credstash_secrets.sh
