+++ server/scripts/run_with_credstash.sh
