+++ server/src/ecosystem.config.js
@@ -2,7 +2,7 @@ module.exports = {
   apps: [{
     name: 'server',
     script: 'run_with_credstash.sh',
-    args: 'node server.js',
+    args: 'node src/server.js',
     env: {
       SERVER_PORT: 8081,
       NODE_ENV: 'development',
