+++ tests/helpers.js
@@ -1,7 +1,7 @@
 import { ClientFunction } from 'testcafe';
 import config from './config';
 
-export * from '../server/testUtils';
+export * from '../server/src/testUtils';
 
 /* global baseUrl */
 export const login = ClientFunction((email, password) => window.fetch(`${baseUrl}/api/users/login`, {
